Create shared Homepage base class.
Remove some unused bits.
